% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.genomecache.R
\name{collapse.genomecache}
\alias{collapse.genomecache}
\title{Reduce genome caches by averaging together loci that are very similar up to some set
tolerance level of a specified criterion}
\usage{
\method{collapse}{genomecache}(original.cache, new.cache, subjects = NULL,
  criterion = c("l2.norm", "max.category"), model = c("additive", "full"),
  proportion.tol = 0.01)
}
\arguments{
\item{original.cache}{The path of the genome cache to be reduced.}

\item{new.cache}{The path of the new genome cache to be created.}

\item{subjects}{DEFAULT: NULL. Allows for the specification of a reduced set of individuals from
the original genome cache to be included in the new one. The default includes all individuals.}

\item{criterion}{DEFAULT: "l2.norm". Option to specify criterion for collapsing loci founder 
probabilities/dosages. "l2.norm" means that max l2 norm or Euclidean distance is used, which means
changes across all categories count. "max.category" means only changes in the max category (founder
dosage or diplotype) are used.}

\item{model}{DEFAULT: "additive". If "additive" is specified, criteria are based on dosages. If 
"full", probabilities of diplotypes are used.}

\item{proportion.tol}{DEFAULT: 0.01. If the maximum criterion value at a pair of loci in a data set 
exceeds this value, the loci are not averaged. When all criterion values are below it, the two loci
get averaged. The scale of the parameter is in terms of a proportion, regardless of criteria.}
}
\description{
This function takes an inpute genome cache directory, and produces a smaller version
with similar information. It is particularly useful for large, dense genome caches, with
redundant loci.
}
\examples{
collapse.genomecache()
}
