% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.DOQTL.to.HAPPY.R
\name{convert.DOQTL.to.HAPPY}
\alias{convert.DOQTL.to.HAPPY}
\title{Takes DO-QTL founder haplotype reconstruction output files and re-formats into a HAPPY genome cache}
\usage{
convert.DOQTL.to.HAPPY(DOQTL.recon.output.path, map,
  physical_dist.is.Mb = TRUE, map.locus_name.colname = "SNP_ID",
  map.chr.colname = "Chr", map.physical_dist.colname = "Mb_NCBI38",
  map.genetic_dist.colname = "cM", HAPPY.output.path,
  allele.labels = NULL, chr = c(1:19, "X"), in.log.scale = FALSE,
  founder.probs.Rdata.available = TRUE, samples = NULL,
  simple.alleles = NULL)
}
\arguments{
\item{DOQTL.recon.output.path}{The path to the directory containing DO-QTL founder haplotype output files.}

\item{map}{The map file (which contains important information on the loci) loaded into R.}

\item{physical_dist.is.Mb}{DEFAULT: TRUE. IF true, Mb will be converted to bp within the function.}

\item{map.locus_name.colname}{DEFAULT: "SNP_ID". The column name in the map data that corresponds to locus/marker names.}

\item{map.chr.colname}{DEFAULT: "Chr". The column name in the map data that corresponds to chromosome.}

\item{map.physical_dist.colname}{DEFAULT: "Mb_NCBI38". The column name in the map data that corresponds to physical position.}

\item{map.genetic_dist.colname}{DEFAULT: "cM". The column name in the map data that corresponds to genetic position.}

\item{HAPPY.output.path}{The path to a directory that will be created as the HAPPY-format genome cache.}

\item{allele.labels}{DEFAULT: NULL. Allows for specification of founder labels different from what is in the DO-QTL
output. The DEFAULT of NULL leads to using the labels from the DO-QTL output.}

\item{chr}{DEFAULT: c(1:19, "X"). Allows for specification of the chromosomes. DEFAULT is all the chromosomes from the mouse.}

\item{in.log.scale}{DEFAULT: FALSE. If TRUE, probabilities are assumed to be log(p-value). Output in genomecache is p-value.}

\item{samples}{DEFAULT: NULL. If founder.probs.Rdata is not available, input a vector with the DO identifiers. Can be pulled from 
individual-level files with naming scheme [sample].genotype.probs.Rdata.}

\item{simple.alleles}{DEFAULT: NULL. If founder.probs.Rdata is not available, input a vector with the simple allele labels. Standard is
LETTERS[1:8].}
}
\description{
This function produces a HAPPY-format genome cache from DO-QTL founder haplotype reconstruction output files.
The main output files of importance are the individual-level files with naming scheme [sample].genotype.probs.Rdata.
}
\examples{
convert.DOQTL.to.HAPPY()
}
