% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.qtl2.to.miqtl.R
\name{convert.qtl2.to.HAPPY}
\alias{convert.qtl2.to.HAPPY}
\title{Takes qtl2geno founder haplotype reconstruction output files and re-formats into a HAPPY genome cache}
\usage{
convert.qtl2.to.HAPPY(qtl2.object, cross.object, HAPPY.output.path,
  allele.labels = LETTERS[1:8], chr = c(1:19, "X"),
  diplotype.order = c("qtl2", "DOQTL", "CC"))
}
\arguments{
\item{qtl2.object}{The full probabilities output as a list of 3D arrays from qtl2geno.}

\item{cross.object}{The cross object that includes map information as well as phenotype and covariate information.}

\item{HAPPY.output.path}{The path to a directory that will be created as the HAPPY-format genome cache.}

\item{allele.labels}{DEFAULT: NULL. Allows for specification of founder labels different from what is in the DO-QTL
output. The DEFAULT of NULL leads to using the labels from the DO-QTL output.}

\item{chr}{DEFAULT: c(1:19, "X"). Specifies the chromosomes to include in the cache.}

\item{diplotype.order}{DEFAULT: "qtl2". Specifies the order the diplotype columns are in. The wrong order will result
in incorrect rotation of probabilities into dosages.}
}
\description{
This function produces a HAPPY-format genome cache from qtl2geno founder haplotype reconstruction output files.
}
\examples{
convert.qtl2.to.HAPPY()
}
