% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.mediation.scan.R
\name{extract.chr.max.statistics.from.genomewide}
\alias{extract.chr.max.statistics.from.genomewide}
\title{Extracts chromosome-level statistics from the full output of genome-wide permutation scans, allowing chromosome-wide and genome-wide significance
to be determined from the same output.}
\usage{
\method{extract}{chr.max.statistics.from.genomewide}(full.perm.scans, chr, use.lod = FALSE)
}
\arguments{
\item{full.perm.scans}{Output object from generate.qr.permutation.index.matrix().}

\item{chr}{The chromosome that will have its association scores extracted from the full.perm.scans, essentially
creating permutation scans specific to the chromosome.}

\item{use.lod}{DEFAULT: FALSE. If TRUE, LOD scores are recorded. If FALSE, p-values are recorded, either based on the likelihood ratio test or ANOVA.}
}
\description{
This function extracts the associations for a specified chromosome from genome-wide permutation scans. This allows both genome-wide and chromosome-wide
thresholds to be determined from the same output.
}
\examples{
extract.chr.max.statistics.from.genomewide()
}
