% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.eQTL.scan.R
\name{extract.qr}
\alias{extract.qr}
\title{Pre-compute QR decompositions from genome cache for efficient fixed effect model genome scan.}
\usage{
\method{extract}{qr}(
  genomecache,
  id = "SUBJECT.NAME",
  data,
  formula,
  model = c("additive", "full"),
  condition.loci = NULL,
  chr = "all",
  just.these.loci = NULL,
  use.progress.bar = TRUE
)
}
\arguments{
\item{genomecache}{The path to the genome cache directory. The genome cache is a particularly structured
directory that stores the haplotype probabilities/dosages at each locus. It has an additive model
subdirectory and a full model subdirectory. Each contains subdirectories for each chromosome, which then
store .RData files for the probabilities/dosages of each locus.}

\item{id}{DEFAULT: "SUBJECT.NAME". The is the individual-level ID that is associated with data points 
in the phenotype data. This should be unique for each data point.}

\item{data}{A data frame with outcome and potential covariates. Should also have IDs
that link to IDs in the genome cache, often with the individual-level ID named "SUBJECT.NAME", though others
can be specified with pheno.id.}

\item{formula}{The right side of an lm-style formula with functions of covariates contained in data frame. First
symbol should be "~". Functions of the outcome will be specified in scan.qr().}

\item{model}{DEFAULT: "additive". Specifies how to model the founder haplotype probabilities. The additive options specifies
use of haplotype dosages, and is most commonly used. The full option regresses the phenotype on the actual
diplotype probabilities.}

\item{condition.loci}{DEFAULT: NULL. If loci are specified, they will be included in the null and alternative models.
The loci must be present in the genome cache. Alternatively, conditional scans can be accomplished by regressing out 
the loci effects with condition.out.locus.for.scan(), which does not require a new qr.object.}

\item{chr}{DEFAULT: "all". Specifies which chromosomes to scan.}

\item{just.these.loci}{DEFAULT: NULL. Specifies a reduced set of loci to fit. If loci is just one locus, the alternative model fit
will also be output as fit1.}

\item{use.progress.bar}{DEFAULT: TRUE. Results in a progress bar being displayed.}
}
\description{
This function calculates the QR decomposition for the null model, and all alternative models (per locus)
for a genome scan. This is only possible with a model that does not include any random effects.
}
\examples{
extract.qr()
}
