% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.processes.R
\name{generate.sample.outcomes.matrix}
\alias{generate.sample.outcomes.matrix}
\title{Returns a matrix of outcome samples, either permutations or from the null model of no locus effect}
\usage{
generate.sample.outcomes.matrix(
  scan.object,
  model.type = c("null", "alt"),
  method = c("bootstrap", "permutation", "subsample"),
  subsample.prop = 0.63,
  subsample.chr = NULL,
  use.REML = TRUE,
  use.BLUP = FALSE,
  num.samples,
  seed = 1
)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object.}

\item{model.type}{DEFAULT: "null". "null" specifies sampling processes from the null model. "alt" specifies sampling processes
from the alternative model.}

\item{method}{DEFAULT: "bootstrap". "bootstrap" specifies parametric bootstraps from the given model. "permutation" specifies
parametric permutations that can respect the structure of the data. Permutations are more appropriate if the data have highly
influential data points. "subsample" specifies randomly sampling without replacement some proportion of the data. This is done
by placing NAs in the observations not selected.}

\item{subsample.prop}{DEFAULT: 0.63. The proportion of the original data set to sample.}

\item{subsample.chr}{DEFAULT: NULL. If method "subsample" is specified, no locus need be specified. If NULL, the default
behavior is to grab the locus (and chromosome) with the peak association. If chromosome is specified, its peak locus will
be grabbed to be passed to the scanning procedure.}

\item{use.REML}{DEFAULT: TRUE. Determines whether the variance components for the parametric sampling are 
based on maximizing the likelihood (ML) or the residual likelihood (REML).}

\item{use.BLUP}{DEFAULT: FALSE.This results in the BLUP value of the polgyene effect (assuming a GRM has been given) is used,
rather than sampled. This reduces the variation seen across sampling, which can result in narrower positional confidence 
intervals.}

\item{num.samples}{The number of parametric bootstrap samples to return.}

\item{seed}{DEFAULT: 1. The sampling process is random, thus a seed must be set for samples to be consistent
across machines.}
}
\description{
This function takes an scan.h2lmm() object, and returns a specified number of outcome samples, either permutations or
from the null model of no locus effect.
}
\examples{
generate.sample.outcomes.matrix()
}
