% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{genome.plotter.region}
\alias{genome.plotter.region}
\title{Plot user-specified windows of haplotype-based and snp-based genome scans}
\usage{
genome.plotter.region(haplotype.association = NULL,
  snp.association = NULL, use.lod = FALSE, chr, region.min = NULL,
  region.max = NULL, scale = c("Mb", "cM"), haplotype.col = c("blue",
  "red"), haplotype.lwd = 3, haplotype.lty = 1,
  median.band.col = c("cyan", "pink"), snp.col = c("black", "gray"),
  snp.pch = 20, snp.cex = 0.9, main = "", no.title = FALSE,
  override.title = NULL, y.max.manual = NULL, my.y.line = 2,
  my.y.axis.cex = 1, my.y.lab.cex = 0.5, my.x.line = 2,
  my.x.axis.cex = 1, my.xlab.cex = 1, x.padj = -0.3,
  my.x.labels = TRUE, override.xlab = NULL, include.x.ticks = TRUE,
  drop.x.tick = FALSE, my.title.line = 0.5, my.title.cex = 1,
  hard.thresholds = NULL, thresholds.col = "red",
  thresholds.legend = NULL, thresholds.lwd = 2, use.legend = TRUE,
  my.legend.cex = 0.6, my.legend.pos = "topright", my.bty = "n",
  rug.pos = NULL, rug.col = "gray50")
}
\arguments{
\item{haplotype.association}{DEFAULT: NULL. A list of scan.h2lmm() objects (ROP or multiple imputations). If multiple imputations, median and confidence interval 
on median are plotted. If NULL, presumably only SNP scans will be plotted.}

\item{snp.association}{DEFAULT: NULL. A list of imputed.snp.scan.h2lmm() objects. If NULL, presumably only haplotype-based scans will be plotted.}

\item{use.lod}{DEFAULT: FALSE. Plots either the LOD score or the -log10 p-value.}

\item{chr}{The chromosome to be plotted.}

\item{region.min}{DEFAULT: NULL. The lower bound of the region to be plotted. Should match the scale. If NULL, defaults to the minimum of the specified chromosome.}

\item{region.max}{DEFAULT: NULL. The upper bound of the region to be plotted. Should match the scale. If NULL, defaults to the maximum of the specified chromosome.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM can be used.}

\item{haplotype.col}{DEFAULT: "black". The color of the haplotype-based association score to be plotted.}

\item{median.band.col}{DEFAULT: "gray88". The color of the 95\% confident band plotted around the median.}

\item{snp.col}{DEFAULT: "black". The color of the SNP association score to be plotted.}

\item{main}{DEFAULT: "". Adds a title above the model.}

\item{no.title}{DEFAULT: FALSE. If TRUE, no title is printed.}

\item{override.title}{DEFAULT: NULL. If a string is specified, it is included on plot without any of the default automated title.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,
no threshols legend is used.}

\item{use.legend}{DEFAULT: TRUE. Include a legend for the different associations. If TRUE, the labels are the names of the non.mi.scan.list object.}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{my.legend.pos}{DEFAULT: "topright". Specifies where to put the legend, if specified in use.legend.}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}
}
\description{
This function takes genome scan association outputs and plots the portion that corresponds to a region of a single chromosome.
When multiple imputations are used, includes the 95\% confidence band on the median.
}
\examples{
genome.plotter.region()
}
