% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{genome.plotter.to.pdf}
\alias{genome.plotter.to.pdf}
\title{Plot whole genome and single chromosome windows of haplotype-based genome scan in a PDF output document}
\usage{
genome.plotter.to.pdf(scan.object, chr = c(1:19, "X"), use.lod = FALSE,
  scale = c("Mb", "cM"), main.col = "black", median.band.col = "gray88",
  main = "", y.max.manual = NULL, hard.thresholds = NULL,
  thresholds.col = "red", thresholds.legend = NULL, pdf.output.path,
  pdf.height = 5, pdf.width = 9, ...)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object (ROP or multiple imputations). If multiple imputations, median and confidence interval 
on median are plotted.}

\item{chr}{DEFAULT: c(1:19, "X"). The chromosomes to be plotted. DEFAULT is all the mouse chromosomes.}

\item{use.lod}{DEFAULT: FALSE. Plots either the LOD score or the -log10 p-value.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{median.band.col}{DEFAULT: "gray88". The color of the 95\% confident band plotted around the median.}

\item{main}{DEFAULT: "". Adds a title above the model.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,
no threshols legend is used.}

\item{pdf.output.path}{That path of the PDF file to be generated.}

\item{pdf.height}{DEFAULT: 5. The height of an individual pages of the PDF.}

\item{pdf.width}{DEFAULT: 9. The width of an individual pages of the PDF.}

\item{main.colors}{DEFAULT: "black". The color of the main association score to be plotted.}
}
\description{
This function takes the genome scan output from scan.h2lmm() and plots the whole genome and single chromosome zoom-ins for 
all the specified chromosomes. When multiple imputations are used, includes the 95\% confidence band on the median in the zoomed-in
plots.
}
\examples{
genome.plotter.to.pdf()
}
