% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{grab.locus.from.scan}
\alias{grab.locus.from.scan}
\title{Pulls loci from scan objects based on specified criteria}
\usage{
grab.locus.from.scan(scan.object, use.lod = FALSE, chr = "all",
  criterion = c("min", "max"), return.value = c("marker", "position", "chr",
  "index"))
}
\arguments{
\item{scan.object}{A scan.h2lmm() object (ROP or multiple imputations). If multiple imputations, median and confidence interval 
on median are plotted.}

\item{use.lod}{DEFAULT: FALSE. Specifies whether loci should be selected based on LOD scores or p-values.}

\item{chr}{DEFAULT: "all". The portion of the scan that loci are being pulled from.}

\item{criterion}{DEFAULT: "min". The criterion by which loci are selected. "min" selects the locus with the minimum statistical score.
"max" selects the locus with the maximum statistical score.}

\item{return.value}{DEFAULT: "marker". If "marker", returns the marker name. If "positions", returns the position in both cM and Mb.
If "chr", returns the chr of the peak. If "index", returns the index of the vector.}
}
\description{
This function parses a scan object and returns loci, for instance, the locus with the minimum
p-value.
}
\examples{
grab.locus.from.scan()
}
