% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.mediation.scan.R
\name{mediation.scan.qr}
\alias{mediation.scan.qr}
\title{Runs genome-wide mediation scan from a QR decomposition object based on a specified QTL and phenotype data file.}
\usage{
mediation.scan.qr(
  mediation.qr.object,
  data,
  phenotype,
  chr = "all",
  id = "SUBJECT.NAME",
  debug.single.fit = FALSE,
  use.progress.bar = TRUE,
  ...
)
}
\arguments{
\item{mediation.qr.object}{Output object from extract.mediation.qr()}

\item{data}{A data frame with outcome. Should also have IDs
that link to IDs in the genome cache, often with the individual-level ID named "SUBJECT.NAME", though others
can be specified with pheno.id.}

\item{phenotype}{The column name (or function of column name) of a variable in data. This will become the outcome
of the mediation scan.}

\item{chr}{DEFAULT: "all". Specifies which chromosomes to scan.}

\item{id}{DEFAULT: "SUBJECT.NAME". The is the individual-level ID that is associated with data points 
in the phenotype data. This should be unique for each data point.}

\item{debug.single.fit}{DEFAULT: FALSE. If TRUE, a browser() call is activated after the first locus is fit. This option
allows developers to more easily debug while still using the actual R package.}

\item{use.progress.bar}{DEFAULT: TRUE. Results in a progress bar being displayed.}
}
\description{
This function runs the genome scan based on a QR decomposition object and phenotype data file. This functionality only
works for fixed effect models. Currently written for chromatin mediation of local-eQTL, though could easily be 
generalized to any pairing of mediator and outcome.
}
\examples{
mediation.scan.qr()
}
