% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_effects_plot.R
\name{plot_locus.effect.from.scan}
\alias{plot_locus.effect.from.scan}
\title{Plot the regression coefficients as allele effects for a single locus as a caterpillar plot from a scan object}
\usage{
plot_locus.effect.from.scan(
  scan.object,
  locus,
  main = "",
  names = c("ACI", "BN", "BUF", "F344", "M520", "MR", "WKY", "WN"),
  col = c(rgb(240, 240, 0, maxColorValue = 255), rgb(128, 128, 128, maxColorValue =
    255), rgb(240, 128, 128, maxColorValue = 255), rgb(16, 16, 240, maxColorValue = 255),
    rgb(0, 160, 240, maxColorValue = 255), rgb(0, 160, 0, maxColorValue = 255), rgb(240,
    0, 0, maxColorValue = 255), rgb(144, 0, 224, maxColorValue = 255)),
  add.connect.line = TRUE,
  connect.lwd = 2,
  connect.lty = 1,
  connect.col = "black",
  add = FALSE,
  ...
)
}
\arguments{
\item{scan.object}{A scan object from scan.h2lmm(). Can work for ROP or multiple imputation.}

\item{locus}{The locus for which effect estimates are need. The locus must be included in the scan.}

\item{main}{DEFAULT: "". The title to included in the plot.}

\item{names}{DEFAULT: c("ACI", "BN", "BUF", "F344", "M520", "MR", "WKY", "WN"). These are the strains used in the HS rats.}

\item{col}{DEFAULT: c(rgb(240, 240, 0, maxColorValue=255), rgb(128, 128, 128, maxColorValue=255), rgb(240, 128, 128, maxColorValue=255), 
rgb(16, 16, 240, maxColorValue=255), rgb(0, 160, 240, maxColorValue=255), rgb(0, 160, 0, maxColorValue=255), rgb(240, 0, 0, maxColorValue=255), 
rgb(144, 0, 224, maxColorValue=255)). These are the established Collaborative Cross colors.}

\item{add.connect.line}{DEFAULT: TRUE. Adds a connection line between the effects (or their mean).}

\item{connect.lwd}{DEFAULT: 2. The line width of the connection line, if it is included}

\item{connect.col}{DEFAULT: "black". The color of the connection line, if it is included}

\item{add}{DEFAULT: FALSE. If TRUE, a new plot is not initiated. The purpose is to allow multiple sets of
effects to be included on a single plot.}

\item{connect.ltt}{DEFAULT: 1. The line type of the connection line, if it is included}
}
\description{
This function takes a scan object from scan.h2lmm() and plots out allele effect estimates
based on regression coefficients for a single locus. If multiple imputation was used, these correspond to 
confidence intervals on the mean effect over imputations.
}
\examples{
plot_locus.effect.from.scan()
}
