% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{rint}
\alias{rint}
\title{Returns the rank-based inverse normal transformation}
\usage{
rint(phenotype, prop = 0.5)
}
\arguments{
\item{phenotype}{A vector of phenotype values for which the rank-based inverse normal transformation is output.}

\item{prop}{DEFAULT: 0.5. This allows Inf to not be returned for the maximum of phenotype.}
}
\description{
This function takes a phenotype vector and returns the rank-based inverse normal transformation.
}
\examples{
rint()
}
