% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.processes.R
\name{run.positional.scans}
\alias{run.positional.scans}
\title{Run single chromosome scans on parametric bootstrap samples from the alternative model of a particular locus}
\usage{
run.positional.scans(sim.object, keep.full.scans = TRUE, genomecache, data,
  model = c("additive", "full"), use.par = "h2", use.multi.impute = TRUE,
  num.imp = 11, brute = TRUE, use.fix.par = FALSE, scan.seed = 1,
  do.augment = FALSE, use.augment.weights = FALSE, use.full.null = FALSE,
  added.data.points = 1, ...)
}
\arguments{
\item{sim.object}{A sample object. This is primarily a matrix of parametric bootstrap
samples from a locus.}

\item{keep.full.scans}{DEFAULT: TRUE. If TRUE, the full scans from each sample are kept. This allows for them
to be plotted out, but also increases the memory or storage needed.}

\item{genomecache}{The path to the genome cache directory. The genome cache is a particularly structured
directory that stores the haplotype probabilities/dosages at each locus. It has an additive model
subdirectory and a full model subdirectory. Each contains subdirectories for each chromosome, which then
store .RData files for the probabilities/dosages of each locus.}

\item{data}{A data frame with outcome and potential covariates. Should also have IDs
that link to IDs in the genome cache, often the individual-level ID named "SUBJECT.NAME".}

\item{model}{DEFAULT: additive. Specifies how to model the founder haplotype probabilities. The additive options specifies
use of haplotype dosages, and is most commonly used. The full option regresses the phenotype on the actual
diplotype probabilities.}

\item{use.par}{DEFAULT: "h2". The parameterization of the likelihood to be used.}

\item{use.multi.impute}{DEFAULT: TRUE. If TRUE, use multiple imputations of genetic data. If FALSE, use ROP.}

\item{num.imp}{DEFAULT: 11. If multiple imputations are used, this specifies the number of imputations to perform.}

\item{brute}{DEFAULT: TRUE. During the optimization to find maximum likelihood parameters, this specifies checking the
boundaries of h2=0 and h2=1. Slightly less efficient, but otherwise the optimization procedure will not directly check
these values.}

\item{use.fix.par}{DEFAULT: TRUE. This specifies an approximate fitting of mixed effect model (Kang et al. 2009). Much
more efficient, as the optimization of h2 only needs to be performed once for the null model rather than every locus. 
Technically less powerful, though in practice it has proven to be almost equal to the exact procedure.}

\item{scan.seed}{DEFAULT: 1. If imputations are used, the sampling procedure is a random process. Specifying a seed allows 
consistent results across runs and machines. This is reset over scans, so that the same imputations are used for each scan.}

\item{do.augment}{DEFAULT: FALSE. Augments the data with null observations for genotype groups. This is an approximately Bayesian 
approach to applying a prior to the data, and can help control highly influential data points.}

\item{use.augment.weights}{DEFAULT: FALSE. Specify non-equal weights on the augmented data points. This allows for the inclusion of
augmented data points to all genotype classes while reducing their overall contribution to the data.}

\item{use.full.null}{DEFAULT: FALSE. Draws augmented data points from the null model. This allows for the inclusion of null data points
that do not influence the estimation of other model parameters as much.}

\item{added.data.points}{DEFAULT: 1. If augment weights are being used, this specifies how many data points should be added in total.}

\item{scan.seed}{DEFAULT: 1.}
}
\description{
This function runs single chromosome scans of parametric bootstrap samples from the alternative model of a particular locus. These
association scans can then be used to estimate a confidence interval on QTL position.
}
\examples{
run.positional.scans()
}
