% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.eQTL.scan.R
\name{run.qr.permutation.threshold.scans}
\alias{run.qr.permutation.threshold.scans}
\title{Runs permutation scans for significance thresholds based on an index permutation matrix object, the phenotype data, and
pre-computed QR decompositions for all the models.}
\usage{
run.qr.permutation.threshold.scans(
  perm.ind.matrix,
  qr.object,
  phenotype,
  data,
  keep.full.scans = FALSE,
  scan.index = NULL,
  id = "SUBJECT.NAME",
  chr = "all",
  just.these.loci = NULL,
  use.progress.bar = FALSE,
  report.perm.scans = FALSE,
  ...
)
}
\arguments{
\item{perm.ind.matrix}{Output object from generate.qr.permutation.index.matrix().}

\item{qr.object}{Output object from extract.qr().}

\item{phenotype}{The column name (or function of column name) of a variable in data. This will become the outcome
of the genome scan.}

\item{data}{A data frame with outcome and potential covariates. Should also have IDs
that link to IDs in the genome cache, often with the individual-level ID named "SUBJECT.NAME", though others
can be specified with id.}

\item{keep.full.scans}{DEFAULT: FALSE. If TRUE, all p-values are kept from all loci. If FALSE, only the minimum p-value
is kept from each scan, greatly reducing size of output.}

\item{scan.index}{DEFAULT: NULL. If NULL, all permutations are run. Integer vector can be specified to run just a 
subset of the permutations.}

\item{id}{DEFAULT: "SUBJECT.NAME". This is the individual-level ID that is associated with data points 
in the phenotype data. This should be unique for each data point.}

\item{chr}{DEFAULT: "all". Specifies which chromosomes to scan.}

\item{just.these.loci}{DEFAULT: NULL. DEFAULT: NULL. Specifies a reduced set of loci to fit. If loci is just one locus, the alternative model fit
will also be output as fit1.}

\item{use.progress.bar}{DEFAULT: FALSE. Results in a progress bar while code runs.}

\item{report.perm.scans}{DEFAULT: FALSE. Outputs the completion of permutation scans.}
}
\description{
This function runs permutation scans in order to calculate significance thresholds. Its results are most valid when
done in a fixed effect model setting and observations are exchangeable.
}
\examples{
run.qr.permutation.threshold.scans()
}
