% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.eQTL.scan.R
\name{scan.qr}
\alias{scan.qr}
\title{Runs genome scan from a QR decomposition object and phenotype data file.}
\usage{
scan.qr(qr.object, data, phenotype, return.allele.effects = FALSE,
  chr = "all", id = "SUBJECT.NAME", just.these.loci = NULL,
  debug.single.fit = FALSE, use.progress.bar = TRUE, ...)
}
\arguments{
\item{qr.object}{QR decomposition output from extract.qr().}

\item{data}{A data frame with outcome and potential covariates. Should also have IDs
that link to IDs in the genome cache, often with the individual-level ID named "SUBJECT.NAME", though others
can be specified with id.}

\item{phenotype}{The column name (or function of column name) of a variable in data. This will become the outcome
of the genome scan.}

\item{chr}{DEFAULT: "all". Specifies which chromosomes to scan.}

\item{id}{DEFAULT: "SUBJECT.NAME". This is the individual-level ID that is associated with data points 
in the phenotype data. This should be unique for each data point.}

\item{just.these.loci}{DEFAULT: NULL. Specifies a reduced set of loci to fit. If loci is just one locus, the alternative model fit
will also be output as fit1.}

\item{debug.single.fit}{DEFAULT: FALSE. If TRUE, a browser() call is activated after the first locus is fit. This option
allows developers to more easily debug while still using the actual R package.}

\item{use.progress.bar}{DEFAULT: TRUE. Results in a progress bar while code runs.}
}
\description{
This function runs the genome scan based on a QR decomposition object and phenotype data file. This functionality only
works for fixed effect models.
}
\examples{
scan.qr()
}
