% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{single.chr.plotter.w.ci}
\alias{single.chr.plotter.w.ci}
\title{Plot whole genome and single chromosome windows of haplotype-based genome scan in a PDF output document}
\usage{
single.chr.plotter.w.ci(scan.object, qtl.ci.object, ci.type, scan.type,
  these.col = c("#7BAFD4", "red"), scale = "Mb")
}
\arguments{
\item{scan.object}{A scan.h2lmm() object (ROP or multiple imputations). If multiple imputations, median and confidence interval 
on median are plotted. Expected to the scan of the actual data.}

\item{qtl.ci.object}{A run.positional.scans() object. Should contain single chromosome scans from some form of sampling process, such as a parametric bootstrap.}

\item{ci.type}{Positional confidence interval to be included in the title. Example: "Parametric Bootstrap".}

\item{scan.type}{Scan type to be included in the title. Example: "ROP".}

\item{these.col}{Colors to be used for individual artificial scans.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either "Mb" or "cM" is expected.}
}
\description{
This function takes the genome scan output from scan.h2lmm() and plots the whole genome and single chromosome zoom-ins for 
all the specified chromosomes. When multiple imputations are used, includes the 95\% confidence band on the median in the zoomed-in
plots.
}
\examples{
single.chr.plotter.w.ci()
}
