% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{snp.genome.plotter.w.r2}
\alias{snp.genome.plotter.w.r2}
\title{Plot a single chromosome window of a SNP-based genome scan overlayed with r^2 information}
\usage{
snp.genome.plotter.w.r2(snp.scan, r2.object, scale = "Mb",
  y.max.manual = NULL, title = "", alt.col = NULL, this.cex = 1,
  hard.thresholds = NULL, thresholds.col = "red",
  thresholds.legend = NULL, my.legend.cex = 0.6,
  my.legend.pos = "topleft", r2.bounds = NULL)
}
\arguments{
\item{snp.scan}{An imputed.snp.scan.h2lmm() object.}

\item{r2.object}{A pairwise.cor.snp.scan() object.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{title}{DEFAULT: "". Manually adds a max ylim to the plot. Allows multiple genome scans to easily be on the same scale.}

\item{alt.col}{DEFAULT: NULL. Allows for a custom color vector for individual SNPs.}

\item{this.cex}{DEFAULT: 1. Allows for the adjustment of the cex value for the main plot.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{my.legend.pos}{DEFAULT: "topright". Specified position of the legend on the plot.}

\item{r2.bounds}{DEFAULT: NULL. If NULL, no interval is depicted on the plot. If set to a value in [0,1], will include interval
based on the given r2 on the plot.}
}
\description{
This function takes the outputs from imputed.snp.scan.h2lmm() and pairwise.cor.snp.scan() and plots 
the SNP association for a single chromosome, overlayed with r^2 information for a specified SNP.
}
\examples{
snp.genome.plotter.w.r2()
}
