% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_plotter.R
\name{snp.genome.plotter.whole}
\alias{snp.genome.plotter.whole}
\title{Plot whole genome and single chromosome windows of a SNP-based genome scan}
\usage{
snp.genome.plotter.whole(snp.scan, just.these.chr = NULL,
  point.col = "black", scale = "Mb", distinguish.chr.type = c("color",
  "box"), distinguish.box.col = "gray88", distinguish.snp.col = "gray60",
  y.max.manual = NULL, my.y.line = 2, my.y.axis.cex = 1, title = "",
  override.title = NULL, alt.col = NULL, hard.thresholds = NULL,
  thresholds.col = "red", thresholds.legend = NULL, thresholds.lty = 2,
  thresholds.lwd = 1, my.legend.cex = 0.6, my.legend.pos = "topright",
  my.bty = "n", my.x.labels = TRUE, add.chr.to.label = FALSE,
  axis.cram = TRUE, include.x.axis.line = TRUE)
}
\arguments{
\item{snp.scan}{An imputed.snp.scan.h2lmm() object.}

\item{just.these.chr}{DEFAULT: NULL. The chromosomes to be plotted. NULL leads to all chromosomes being plotted.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{title}{DEFAULT: "". Manually adds a max ylim to the plot. Allows multiple genome scans to easily be on the same scale.}

\item{alt.col}{DEFAULT: NULL. Allows for a custom color vector for individual SNPs.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{my.legend.pos}{DEFAULT: "topright". Specified position of the legend on the plot.}

\item{add.chr.to.label}{DEFAULT: FALSE. If TRUE, adds "Chr" before every chromosome label. If FALSE, "Chr" is added as an axis
label under the y-axis.}

\item{axis.cram}{DEFAULT: TRUE. This makes the plot much more likely to include all chromosome labels. With small plots, this could
lead to overlapping labels.}

\item{include.x.axis.line}{DEFAULT: TRUE. IF TRUE, this option adds an x-axis line with ticks between chromosomes.}
}
\description{
This function takes the genome scan output from imputed.snp.scan.h2lmm() and plots the whole genome or single chromosome zoom-ins.
}
\examples{
snp.genome.plotter.whole()
}
