% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.plotter.R
\name{allele.plotter.whole}
\alias{allele.plotter.whole}
\title{Plot the regression coefficients or BLUPs as allele effects for a genome scan (whole genomes or subset of chromosomes)}
\usage{
allele.plotter.whole(scan.object, just.these.chr = NULL, scale = "Mb",
  imp.confint.alpha = NULL, main.colors = c(rgb(240, 240, 0, maxColorValue =
  255), rgb(128, 128, 128, maxColorValue = 255), rgb(240, 128, 128,
  maxColorValue = 255), rgb(16, 16, 240, maxColorValue = 255), rgb(0, 160, 240,
  maxColorValue = 255), rgb(0, 160, 0, maxColorValue = 255), rgb(240, 0, 0,
  maxColorValue = 255), rgb(144, 0, 224, maxColorValue = 255)),
  use.legend = TRUE, main = "", my.bty = "n", my.lwd = rep(1.25, 8),
  set.plot.limit = c(-10, 10), my.legend.cex = 0.6,
  my.legend.pos = "topright", transparency = 0.6, y.max.manual = NULL,
  y.min.manual = NULL, my.y.line = 2, my.y.axis.cex = 1,
  my.y.lab.cex = 0.9, no.title = FALSE, override.title = NULL,
  add.chr.to.label = FALSE, alternative.labels = NULL, axis.cram = TRUE,
  include.x.axis.line = TRUE)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object that had return.allele.effects=TRUE.}

\item{just.these.chr}{DEFAULT: NULL. Specifies a subset of the chromosomes to be plotted. NULL results in all chromosomes being plotted.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{main.colors}{DEFAULT: CC founder colors. The colors that correspond to the founder alleles.}

\item{use.legend}{DEFAULT: TRUE. Include a legend for the different associations. If TRUE, the labels are the names of the non.mi.scan.list object.}

\item{main}{DEFAULT: NULL. Adds a title above the model.}

\item{my.lwd}{DEFAULT: rep(1.5, 8). Specifies the lwds for the alleles.}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{my.legend.pos}{DEFAULT: "topright". Specifies where to put the legend, if specified in use.legend.}

\item{transparency}{DEFAULT: 0.6. The transparency level for the confidence bands. 0 is invisible, 1 is solid.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value.}

\item{y.min.manual}{DEFAULT: NULL. Manually adds a min y-value.}

\item{no.title}{DEFAULT: FALSE. If TRUE, produces a plot with no title, not even the automatic
title components like formula and sample size.}

\item{override.title}{DEFAULT: NULL. If non-null, overrides automatic title components like
formula and sample size.}

\item{add.chr.to.label}{DEFAULT: FALSE. If TRUE, "chr" precedes integers on the x-axis.}

\item{alternative.labels}{DEFAULT: NULL. If non-null, specifies alternative labels for the
founder alleles.}

\item{axis.cram}{DEFAULT: TRUE. This makes the plot much more likely to include all chromosome labels. With small plots, this could
lead to overlapping labels.}

\item{include.x.axis.line}{DEFAULT: TRUE. IF TRUE, this option adds an x-axis line with ticks between chromosomes.}
}
\description{
This function takes a genome scan object from scan.h2lmm() and plots out allele effect estimates
based on regression coefficients or BLUPs.
}
\examples{
allele.plotter.whole()
}
