% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.DOQTL.to.HAPPY.R
\name{convert.DOQTL.to.HAPPY}
\alias{convert.DOQTL.to.HAPPY}
\title{Takes DO-QTL founder haplotype reconstruction output files and re-formats into a HAPPY genome cache}
\usage{
convert.DOQTL.to.HAPPY(DOQTL.recon.output.path, map.path, HAPPY.output.path,
  allele.labels = NULL, chr = c(1:19, "X"))
}
\arguments{
\item{DOQTL.recon.output.path}{The path to the directory containing DO-QTL founder haplotype output files.}

\item{map.path}{The path to the map file. The map file contains data on the loci. It should be a tab-delimited 
file with columns labeled "marker", "chr", "bp", and "pos". "pos" represents map distance in cM. "bp" should 
be in bp, not Mb.}

\item{HAPPY.output.path}{The path to a directory that will be created as the HAPPY-format genome cache.}

\item{allele.labels}{DEFAULT: NULL. Allows for specification of founder labels different from what is in the DO-QTL
output. The DEFAULT of NULL leads to using the labels from the DO-QTL output.}

\item{chr}{DEFAULT: c(1:19, "X"). Allows for specification of the chromosomes. DEFAULT is all the chromosomes from the mouse.}
}
\description{
This function produces a HAPPY-format genome cache from DO-QTL founder haplotype reconstruction output files.
The main output files of importance are the individual-level files with naming scheme [sample].genotype.probs.Rdata.
}
\examples{
convert.DOQTL.to.HAPPY()
}
