% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{genome.plotter.chr}
\alias{genome.plotter.chr}
\title{Plot single chromosome windows of haplotype-based genome scan}
\usage{
genome.plotter.chr(scan.object, chr, use.lod = FALSE, scale = c("Mb", "cM"),
  main.col = "black", median.band.col = "gray88", main = "",
  y.max.manual = NULL, my.legend.cex = 0.6, hard.thresholds = NULL,
  thresholds.col = "red", thresholds.legend = NULL)
}
\arguments{
\item{scan.object}{A scan.h2lmm() object (ROP or multiple imputations). If multiple imputations, median and confidence interval 
on median are plotted.}

\item{chr}{The chromosome to be plotted.}

\item{use.lod}{DEFAULT: FALSE. Plots either the LOD score or the -log10 p-value.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM can be used.}

\item{median.band.col}{DEFAULT: "gray88". The color of the 95\% confident band plotted around the median.}

\item{main}{DEFAULT: "". Adds a title above the model.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,
no threshols legend is used.}

\item{main.colors}{DEFAULT: "black". The color of the main association score to be plotted.}
}
\description{
This function takes the genome scan output from scan.h2lmm() and plots the portion that corresponds to a single chromosome.
When multiple imputations are used, includes the 95\% confidence band on the median.
}
\examples{
genome.plotter.chr()
}
