% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{genome.plotter.whole}
\alias{genome.plotter.whole}
\title{Plot one or more haplotype-based genome scans flexibly (whole genomes or subset of chromosomes)}
\usage{
genome.plotter.whole(scan.list, use.lod = FALSE, just.these.chr = NULL,
  scale = "Mb", main.colors = c("black", "gray48", "blue"),
  use.legend = TRUE, main = "", my.legend.cex = 0.6,
  my.legend.lwd = NULL, my.legend.pos = "topright", y.max.manual = NULL,
  hard.thresholds = NULL, thresholds.col = "red",
  thresholds.legend = NULL)
}
\arguments{
\item{scan.list}{A list of scan.h2lmm() objects that are to be plotted within the same genome scan plot.}

\item{use.lod}{DEFAULT: FALSE. Plots either the LOD score or the -log10 p-value.}

\item{just.these.chr}{DEFAULT: NULL. Specifies a subset of the chromosomes to be plotted. NULL results in all chromosomes being plotted.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{main.colors}{DEFAULT: "black". The color of the main association score to be plotted.}

\item{use.legend}{DEFAULT: TRUE. Include a legend for the different associations. If TRUE, the labels are the names of the non.mi.scan.list object.}

\item{main}{DEFAULT: NULL. Adds a title above the model.}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{my.legend.lwd}{DEFAULT: NULL. If NULL, all lines have lwd=1.5. If not, option specifies the lwds.}

\item{my.legend.pos}{DEFAULT: "topright". Specifies where to put the legend, if specified in use.legend.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,
no threshols legend is used.}
}
\description{
This function takes the genome scan output from scan.h2lmm() and flexibly plots out the genome scan.
}
\examples{
genome.plotter.whole()
}
