% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plotter.R
\name{snp.genome.plotter.whole}
\alias{snp.genome.plotter.whole}
\title{Plot whole genome and single chromosome windows of a SNP-based genome scan}
\usage{
snp.genome.plotter.whole(snp.scan, just.these.chr = NULL, scale = "Mb",
  y.max.manual = NULL, title = "", alt.col = NULL,
  hard.thresholds = NULL, thresholds.col = "red",
  thresholds.legend = NULL, my.legend.cex = 0.6,
  my.legend.pos = "topright")
}
\arguments{
\item{snp.scan}{An imputed.snp.scan.h2lmm() object.}

\item{just.these.chr}{DEFAULT: NULL. The chromosomes to be plotted. NULL leads to all chromosomes being plotted.}

\item{scale}{DEFAULT: "Mb". Specifies the scale of genomic position to be plotted. Either Mb or cM are expected.}

\item{y.max.manual}{DEFAULT: NULL. Manually adds a max y-value. Allows multiple genome scans to easily be on the same scale.}

\item{title}{DEFAULT: "". Manually adds a max ylim to the plot. Allows multiple genome scans to easily be on the same scale.}

\item{alt.col}{DEFAULT: NULL. Allows for a custom color vector for individual SNPs.}

\item{hard.thresholds}{DEFAULT: NULL. Specify one or more horizontal threshold lines.}

\item{thresholds.col}{DEFAULT: "red". Set the colors of the specified thresholds.}

\item{thresholds.legend}{DEFAULT: NULL. If non-NULL, string arguments used as labels in thresholds legend. If NULL,}

\item{my.legend.cex}{DEFAULT: 0.6. Specifies the size of the text in the legend.}

\item{my.legend.pos}{DEFAULT: "topright". Specified position of the legend on the plot.}
}
\description{
This function takes the genome scan output from imputed.snp.scan.h2lmm() and plots the whole genome or single chromosome zoom-ins.
}
\examples{
snp.genome.plotter.whole()
}
